AES-FF1, AES-FF3-1 genvals and tests

+++ src/crypto/src/NIST.CVP.Crypto.Common/Symmetric/BlockModes/FFX/FfxModeBlockCipherParameters.cs
@@ -10,6 +10,6 @@ namespace NIST.CVP.Crypto.Common.Symmetric.BlockModes.Ffx
         public BitString Key { get; set; }
         public BitString Payload { get; set; }
         public bool UseInverseCipherMode => false;
-        public short Radix { get; set; }
+        public int Radix { get; set; }
     }
 }
