+++ src/crypto/src/NIST.CVP.Crypto.Common/Symmetric/BlockModes/FFX/IFfxModeBlockCipherParameters.cs
@@ -5,6 +5,6 @@ namespace NIST.CVP.Crypto.Common.Symmetric.BlockModes.Ffx
         /// <summary>
         /// The number base.
         /// </summary>
-        short Radix { get; }
+        int Radix { get; }
     }
 }
